requirements.txt
evdev==0.7.0
pycairo==1.16.2
pygobject==3.26.1
